Move HomeViewer and LogInOut
