+++ src/Layout/Layout.jsx
@@ -4,11 +4,11 @@ import * as qs from 'query-string';
 
 import DeviceList from '../components/DeviceList/DeviceList';
 import DeviceViewer from '../components/DeviceViewer/DeviceViewer';
-import HomeViewer from '../HomeViewer/HomeViewer';
+import HomeViewer from '../components/HomeViewer/HomeViewer';
 import ErrorDisplay from '../components/ErrorDisplay/ErrorDisplay';
 import Dashboard from '../components/Dashboard/Dashboard';
 import ModalDialog from '../Modal/Modal';
-import LogInOut from '../LogInOut/LogInOut';
+import LogInOut from '../components/LogInOut/LogInOut';
 
 import './Layout.css';
